cleanup(ε.5): extract decode_block/process_candidates.rs (engine room)
Fifth slice of the `decode_block.rs` per-stage split — the engine
room. Moves the entire "Public entry" section into a sibling
submodule: facade entries + multipass drivers + refine passes +
per-candidate processing loop + OSD-still-inline
`process_one_candidate_inner`.

Items moved (1,587 lines):
- Facade: `decode_block`, `decode_block_tuned`,
  `decode_block_with_ap`, `decode_block_with_ap_tuned`,
  `decode_block_into`, `decode_block_into_tuned`
- Multipass drivers: `decode_block_multipass` (host f32 +
  embedded variants), `recompute_nsync`, `recompute_snr_xsnr2`,
  `xsnr2_db_simple`, `fine_refine_pass1` (host + embedded)
- Refine pass: `RefinedCandidate` alias, `pass1_limit`,
  `PASS1_LIMIT_DEFAULT`, `refine_candidates`,
  `refine_candidates_into`, `refine_candidates_with`
- Per-candidate loop: `sync_quality_block0`, `LlrT` alias,
  `bp_step_select`, `process_candidates`,
  `process_candidates_tuned`, `process_candidates_tuned_with_ap`,
  `process_candidates_into` family (4 variants),
  `process_candidates_with_ap`
- `process_one_candidate_inner` — still the OSD-dispatch host
  (ε.6 lifts the OSD block out into `osd_strategy.rs`)
- `WSJTX_NHARDERRORS_MAX` constant

API surface preserved. Parent re-exports the public entries via
`pub use process_candidates::*`, split into base / fft-rustfft /
fixed-point / doc-hidden groups so cfg-gated items don't leak
across builds. `process_one_candidate_inner` is re-exported as
`pub(in crate::ft8)` so the host `super::decode::*` path keeps
reaching it via the same `mfsk_core::ft8::decode_block::*` route.

`super::llr::*` inline path references in the moved body rewritten
to `super::super::llr::*` to reflect the new depth in the module
tree (`decode_block::process_candidates → ft8::llr` is two
super-hops, not one); the `decode_frame` doc link in the
`decode_block` docstring patched the same way.

Parent cleanup: dropped all imports the parent no longer needs
(super::{decode,llr,message,params,wave_gen}, crate::core::scalar,
crate::core::sync, crate::fec::ldpc::{bp,osd}, alloc::*,
num_traits::Float). Tests now self-import what they need.

Acceptance:
- `cargo check` clean on all 4 feature cells, no warnings
- FT8 qso3_busy golden tests pass (WSJT-X AP-off floor, JTDX
  recall floor, AP-on strict-superset, AP-on subtract diag)
- 7 `ft8::decode_block::tests::*` lib tests pass
- `cargo clippy --features full --lib` clean, `cargo fmt --check` clean
- Diff: decode_block.rs −1566 / +21, new process_candidates.rs +1596

`decode_block.rs` is now down to **416 lines** (started at 3,517 —
**88 % reduction**). Final ε slice: ε.6 lifts the OSD dispatch
block (lines 2958-3016 of the original) out of
`process_one_candidate_inner` into a dedicated `osd_strategy.rs`
so #63 (WSJT-X-faithful npre1 precoding) has a clean hook point.

Stacks on #80 (ε.4). Rebases to main when #80 lands.
